Machine learning, binary classification models
This project revolves around assisting the mobile operator "Megaline" in recommending appropriate tariffs, "Smart" or "Ultra", to its customers based on their behavior data. Utilizing preprocessed customer data, the aim is to develop a classification model achieving a minimum accuracy of 0.75 on the test set. The dataset is divided into training, validation, and test subsets for model evaluation, where different models with varied hyperparameters are explored to optimize accuracy. As an additional task, model adequacy is examined to ensure reliability.